requirements.txt
# для yoloface -- детектора лиц
joblib==1.2.0
matplotlib==3.5.1
numpy==1.22.4
onnx==1.12.0
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.3.0
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
setuptools==61.2.0
#thop==0.1.1.post2207130030
thop
#torch==1.12.1+cu116
torch
#torchvision==0.13.1+cu116
torchvision
#tqdm==4.64.0
tqdm
#wandb==0.13.6
wandb


# для распознования лиц
facenet-pytorch